backend/requirements.txt
SQLAlchemy==1.3.13
Flask-SQLAlchemy==2.4.1
Flask-Mail==0.9.1
psycopg2==2.8.4